feat; Implemented read lists
